Reduce allocations when writing longs
StringBuilder.append(Long) allocates a temporary string, while
StringBuilder.append(long) does not and writes digits directly.

+++ src/main/java/com/timgroup/statsd/NonBlockingStatsDClient.java
@@ -615,7 +615,7 @@ public class NonBlockingStatsDClient implements StatsDClient {
 
             sendMetric(new StatsDMessage<Long>(aspect, type, value, sampleRate, timestamp, tags) {
                 @Override protected void writeValue(StringBuilder builder) {
-                    builder.append(this.value);
+                    builder.append(this.value.longValue());
                 }
             });
         }
